# src/main/java/com/bootcamp/clientms/dto/response/ClientResponse.java
package com.bootcamp.clientms.dto.response;

import com.bootcamp.clientms.domain.Client;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class ClientResponse {
  private String id;
  private String firstName;
  private String lastName;
  private String dni;
  private String email;

  public static ClientResponse from(Client client) {
    return new ClientResponse(
            client.getId(),
            client.getFirstName(),
            client.getLastName(),
            client.getDni(),
            client.getEmail()
    );
  }
}